Tests: Move intercepting delivery class into separate file for potential future use
